+++ src/attr/_funcs.py
@@ -6,29 +6,31 @@ from ._compat import iteritems
 from ._make import Attribute, NOTHING, fields
 
 
-def asdict(inst, recurse=True, filter=None, dict_factory=dict):
+def asdict(inst, recurse=True, filter=None, dict_factory=dict,
+           retain_collection_types=False):
     """
-    Return the ``attrs`` attribute values of *inst* as a dict. Optionally
-    recurse into other ``attrs``-decorated classes.
+    Return the ``attrs`` attribute values of *inst* as a dict.
+
+    Optionally recurse into other ``attrs``-decorated classes.
 
     :param inst: Instance of an ``attrs``-decorated class.
-
     :param bool recurse: Recurse into classes that are also
         ``attrs``-decorated.
-
     :param callable filter: A callable whose return code deteremines whether an
         attribute or element is included (``True``) or dropped (``False``).  Is
         called with the :class:`attr.Attribute` as the first argument and the
         value as the second argument.
-
-    :param callable dict_factory: A callable to produce dictionaries from. For
+    :param callable dict_factory: A callable to produce dictionaries from.  For
         example, to produce ordered dictionaries instead of normal Python
         dictionaries, pass in ``collections.OrderedDict``.
+    :param bool retain_collection_types: Do not convert to ``list`` when
+        encountering an attribute which is type ``tuple`` or ``set``.  Only
+        meaningful if ``recurse`` is ``True``.
 
-    :rtype: :class:`dict`
+    :rtype: return type of *dict_factory*
 
-    .. versionadded:: 16.0.0
-        *dict_factory*
+    ..  versionadded:: 16.0.0 *dict_factory*
+    ..  versionadded:: 16.1.0 *retain_collection_types*
     """
     attrs = fields(inst.__class__)
     rv = dict_factory()
@@ -41,12 +43,13 @@ def asdict(inst, recurse=True, filter=None, dict_factory=dict):
                 rv[a.name] = asdict(v, recurse=True, filter=filter,
                                     dict_factory=dict_factory)
             elif isinstance(v, (tuple, list, set)):
-                rv[a.name] = [
+                cf = v.__class__ if retain_collection_types is True else list
+                rv[a.name] = cf([
                     asdict(i, recurse=True, filter=filter,
                            dict_factory=dict_factory)
                     if has(i.__class__) else i
                     for i in v
-                ]
+                ])
             elif isinstance(v, dict):
                 df = dict_factory
                 rv[a.name] = df((